% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-arrange.R
\name{gg_arrange}
\alias{gg_arrange}
\title{Arrange Multiple ggplots}
\usage{
gg_arrange(
  ...,
  plotlist = NULL,
  ncol = NULL,
  nrow = NULL,
  tag.levels = NULL,
  tag.bold = FALSE,
  tag.size = NULL,
  align = c("none", "h", "v", "hv"),
  widths = 1,
  heights = 1,
  plot.margin = NULL,
  legend.position = "right",
  legend.collect = FALSE,
  legend.grob = NULL
)
}
\arguments{
\item{...}{List of plots to be arranged into the grid.}

\item{plotlist}{(optional) list of plots to display.}

\item{ncol}{(optional) number of columns in the plot grid.}

\item{nrow}{(optional) number of rows in the plot grid.}

\item{tag.levels}{(optional) list of labels to be added to the plots. You
can also set "A", "a", "1", "[A]", "[a]", "[1]", "(A)", "(a)"or "(1)" to
auto-generate tag sequence. In addition, you can also directly assign values
through the character vector.}

\item{tag.bold}{logical value, whether to make tags bold.}

\item{tag.size}{tags size.}

\item{align}{(optional) Specifies whether graphs in the grid should be
horizontally ("h") or vertically ("v") aligned. Options are "none" (default),
"hv" (align in both directions), "h", and "v".}

\item{widths}{(optional) numerical vector of relative columns widths. For example,
in a two-column grid, widths = c(2, 1) would make the first column twice as
wide as the second column.}

\item{heights}{same as widths but for column heights.}

\item{plot.margin}{margin around entire plot (vector with the sizes of the top,
right, bottom, and left margins).}

\item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector)}

\item{legend.collect}{logical value. Default is FALSE. If TRUE, a common
unique legend will be created for arranged plots.}

\item{legend.grob}{a legend grob as returned by the function get_legend().
If provided, it will be used as the common legend.}
}
\value{
return an object of class ggarrange, which is a ggplot or a list of ggplot.
}
\description{
Arrange multiple ggplots on the same page. Wrapper around plot_grid().
Can also create a common unique legend for multiple plots.
}
