% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-density.R
\name{gg_density}
\alias{gg_density}
\title{Density plot}
\usage{
gg_density(
  data,
  x,
  y = "density",
  group = NULL,
  facet = NULL,
  x.breaks = NULL,
  y.breaks = NULL,
  legend.title = waiver(),
  legend.label = waiver(),
  legend.position = waiver(),
  add = c("mean", "median", "none"),
  palate = NULL,
  theme = theme_sci(),
  ...
)
}
\arguments{
\item{data}{data}

\item{x}{x}

\item{y}{y}

\item{group}{group}

\item{facet}{facet}

\item{x.breaks}{x breaks.}

\item{y.breaks}{y breaks.}

\item{legend.title}{legend title.}

\item{legend.label}{legend labels}

\item{legend.position}{legend position.}

\item{add}{allowed values are one of "mean" or "median" (for adding mean or
median line, respectively).}

\item{palate}{the color palette to be used for coloring or filling by groups.
Allowed values include "grey" for grey color palettes; brewer palettes e.g.
"RdBu", "Blues", ...; or custom color palette e.g. c("blue", "red"); and
scientific journal palettes from ggsci R package, e.g.: "npg", "aaas",
"lancet", "jco", "ucscgb", "uchicago", "simpsons" and "rickandmorty".}

\item{theme}{function, ggplot2 theme name. Default value is theme_sci().
Allowed values include ggplot2 official themes: theme_gray(), theme_bw(),
theme_minimal(), theme_classic(), theme_void(), ....}

\item{...}{other arguments to be passed to geom_density}
}
\value{
ggplot
}
\description{
Density plot
}
