% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{theme_sci}
\alias{theme_sci}
\title{ggplo2 theme for SCI}
\usage{
theme_sci(
  font.size = 12,
  font.family = "serif",
  line.size = 0.25,
  legend.key.size = 0.8,
  face.bold = FALSE,
  panel.grid.major = FALSE,
  panel.grid.minor = FALSE,
  panel.border = FALSE,
  panel.spacing = 0.6,
  strip.background = "gray90",
  aspect.ratio = NULL,
  ...
)
}
\arguments{
\item{font.size}{font size, default 12,}

\item{font.family}{font family, detault Times New Roman.}

\item{line.size}{line size, default 0.25.}

\item{legend.key.size}{legend key size, dsfault 0.8.}

\item{face.bold}{a logical, whether bold the title of axis, plot, strip, and legend, default FALSE.}

\item{panel.grid.major}{panel grid major.}

\item{panel.grid.minor}{panel grid minor.}

\item{panel.border}{a logical, whether plot panel border, default FALSE.}

\item{panel.spacing}{panel spacing.}

\item{strip.background}{strip background.}

\item{aspect.ratio}{the ratio of the width to the height, default NULL.}

\item{...}{further arguments pass to the [ggplot2::theme] function.}
}
\description{
ggplo2 theme for SCI
}
